adeline-brain/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# LangGraph multi-agent orchestration
langgraph==0.1.5
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
openai==1.30.1

# Database — Hippocampus (pgvector)
pgvector==0.2.5
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
prisma==0.13.1

# GraphRAG
neo4j==5.20.0

# Auth sessions + rate-limiting
redis==5.0.4
fastapi-limiter==0.1.6

# Utilities
python-dotenv==1.0.1
httpx==0.27.0

# Researcher tool — auto primary-source search
tavily-python==0.3.9
beautifulsoup4==4.12.3
numpy==1.26.4

# PDF generation
reportlab==4.1.0
